[OneExplorer] Separate ArtifactLocator.ts
This commit separate Artifact, ArtifactAttr, Locator and LocatorRunner classes
from ConfigObject.ts

ONE-vscode-DCO-1.0-Signed-off-by: dayo09 <dayoung.lee@samsung.com>
